requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.25.0
openpyxl>=3.1.0
chardet>=5.2.0
thefuzz>=0.22.0
python-Levenshtein>=0.21.0
python-dateutil>=2.8.0
plotly>=5.18.0
python-dotenv>=1.0.0
pydantic>=2.5.0
azure-ai-documentintelligence>=1.0.0
azure-identity>=1.15.0
xmltodict>=0.13.0
pyarrow>=14.0.0
openai>=1.0.0
azure-ai-projects>=1.0.0b1
mcp[cli]>=1.0.0
scikit-learn>=1.3.0
